pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unist"
version = "0.1.0"
description = "A Python package for UniST: A Unified Computational Framework for 3D Spatial Transcriptomics Reconstruction"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Lan Shui", email = "lshui@mdanderson.org"}
]
keywords = ["3D Spatial Transcriptomics", "Generative AI", "Point Cloud Upsampling", "Implicit Neural Representations", "Slice Interpolation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "torch>=1.8.0",
    "einops==0.7.0",
    "numpy>=1.19.0",
    "scikit-learn==1.4.2",
    "scipy>=1.5.0",
    "pandas>=1.0.0",
    "anndata>=0.8.0",
    "open3d>=0.13.0",
    "tqdm",
    "h5py",
    "ninja",
    "tensorflow<2.11",
    "imageio",
    "natsort",
    "pillow",
    "tifffile",
    "pyvista>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[tool.setuptools]
packages = [
    "downstream",
    "downstream.morph",
    "downstream.vis",
    "interpolation",
    "interpolation.ReHo",
    "metrics",
    "upsampling",
]

[tool.setuptools.package-data]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
